Update patch set 2
Patch Set 2:

> Why so many empty files in prototypes/xci/file/? I can't
 > immediately understand why.

I didn't want to put more effort without you reviewing it and providing feedback. They are just placeholders to show what I intend to do.

So, they will be similar to what we have in prototypes/openstack-ansible but with different content per flavor.

    https://gerrit.opnfv.org/gerrit/gitweb?p=releng.git;a=blob;f=prototypes/openstack-ansible/playbooks/inventory
    https://gerrit.opnfv.org/gerrit/gitweb?p=releng.git;a=blob;f=prototypes/openstack-ansible/playbooks/configure-xcimaster.yml
    https://gerrit.opnfv.org/gerrit/gitweb?p=releng.git;a=blob;f=prototypes/openstack-ansible/var/ubuntu.yml

They will be the ones used when the actual deployment is done with openstack-ansible based on the flavor chosen by developer (or CI).

Patch-set: 2
Label: Verified=0