xci: Add ansible files for sandbox mini flavor
This patch adds mini target host configuration playbook, inventory, variable
files and so on to use until we have time to fix things more properly and some
other functionality becomes available in upstream. This approach will result in
duplicates but it is something we can perhaps live with.

The user variables and openstack user configuration will be different between
flavors and keeping deployment specific configuration.

Change-Id: Ie312dfa2047d094b590de970c1436d69f4c69668
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>