xci: Add ansible files for sandbox aio flavor
This patch adds aio opnfv host configuration playbook, inventory, variable
files and so on to use until we have time to fix things more properly and some
other functionality becomes available in upstream. This approach will result in
duplicates but it is something we can perhaps live with.

The playbook to configure opnfv host is different from the playbook used for
the rest of the flavors as our target host is the opnfv host itself. This is
the reason for not having target host configuration playbook.

This installation of aio will be driven by upstream scripts/playbooks entirely
and we just execute the scripts we have in openstack-ansible repo.

Change-Id: Ica814bdac1d324414f3add382a8241de49c93a2e
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>