Update patch set 1
Patch Set 1: Code-Review+1

(8 comments)

Well, overall it's OK but I am curious to see how this is going to be used. However, I am somewhat nervous that everything is built around Ubuntu so it will make it harder to spot all the ubuntu-specific cases in the future when more distributions will be added. If we can be proactive here that would be nice.

Patch-set: 1
Label: Code-Review=+1