Update patch set 2
Patch Set 2:

I agree with Markos, currently there is only one "disable" entry in the playbook, but shall be better to filter by the haproxy setting as well. Also, if  we feel we need the ability to configure it, shouldn't be better to send a patch to OSA to make this var configurable?

Patch-set: 2
Reviewer: Gerrit User 4057 <4057@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0