xci: Create periodic osa jobs to find working version
These jobs run periodically against the HEAD of master and stable/ocata
branches in order to attempt deployment and help us to find more
recent working versions to pin.

In perfect world, these jobs should be run post-merge, helping us
to move faster but lack of resources prevent us doing that.

One thing to note here is that we use the pinned version of bifrost
in order to make sure we change one thing at a time to reduce the
delta.

Change-Id: Ieaff7e80581c7577b253b3646ca386313295a45a
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>