change openrc file path to /etc/yardstick/openstack.creds
JIRA: YARDSTICK-623

Now almost yardstick openrc file is in /etc/yardstick/openstack.creds.
But only in prepare_env.sh it is /home/opnfv/openrc.
So I unified it and change it to /etc/yardstick/openstack.creds in the
follow patch:
https://gerrit.opnfv.org/gerrit/#/c/32005/
Also there is some adapt work to do in releng.

Change-Id: I7764582dec1d9a8f74e53b23e23d98021a20589b
Signed-off-by: chenjiankun <chenjiankun1@huawei.com>