Trigger docker builder when a new tag pushed
Currently after the PTL push a new tag in the branch, the PTL must
trigger the docker build job manually in Jenkins. This may cause some
issue:
- two seperate steps
- PTL may not have the Jenkins priority
- PTL should also be careful for the tag name, for example:
    * tag in repo is danube.1.0
    * RELEASE_VERSION is set to 1.0

In this patch, when the PTL pushes a new tag in repo, the build job will
be triggered automatically and set the RELEASE_VERSION from the tag in the
repo.

Change-Id: I04fc541be092ba82a26b0a99034eceebc9dd8258
Signed-off-by: Julien <zhang.jun3g@zte.com.cn>