Update patch set 4
Patch Set 4:

XCI_MODE or XCI_RUN doesn't really match to what the term XCI_LOOP tries to tell in my opinion.

To me loops in CI are the activities that can be run in parallel or in sequence and as independently as possible from each other to provide feedback about the certain aspects of the thing we are building.

In XCI context, the loops I named in this patch aim to reflect this, providing feedback to us about the state of things.

Apart from this, the term loop has also been in use within OPNFV in CI and Test contexts.

https://wiki.opnfv.org/dosearchsite.action?queryString=loop

Patch-set: 4