Update patch set 1
Patch Set 1:

> I think this is not enough since the test jobs are still part of
 > virtual scenario upstream jobs.
 > https://gerrit.opnfv.org/gerrit/gitweb?p=releng.git;a=blob;f=jjb/armband/armband-ci-jobs.yml;hb=HEAD#l155
 > 
 > this will cause failures for the upstream jobs with the error
 > "none-existent jobs"
 > 
 > If you don't care about the failures, it is fine but I would like
 > them not to be red.
Thanks for pointing this out Fatih. Indeed I missed this aspect of it. I don't want to see the failed jobs, but I'm not sure how to fix it at this point; we want to deploy the virtual pod, but not run tests on it. I'll find a way to figure it out

Patch-set: 1
Reviewer: Gerrit User 5267 <5267@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0