Revert "Cleanup dirty daisy build workspace directory"
This reverts commit 3ede21905ba5df94df06952e90d89ec0c16e7bb7.

The deleted line in daisy4nfv-basic.sh in this PS was used exclusively
by a job to clean up files with wrong ownership in workspace previously.
Now, daisy4nfv-basic.sh is used combind with build/test scripts in job,
so we need to delete this line, otherwise, we encountered workspace was
wiped which resulted in build/test scripts failed to run.

Change-Id: I8812a495db789838c3a88510a05956a11e8227d5
Signed-off-by: Zhijiang Hu <hu.zhijiang@zte.com.cn>