xci: Add role to create libvirt vms
This role creates the vms for the chosen flavor.

The vms have 5 interfaces; 1 per network
- admin
- management
- vxlan
- storage
- vlan

This is part of series of changes to separate VM node creation from
bifrost node enrollment/deployment in order to ensure we have ability
to configure libvirt vms and networks properly.

The vm.yml file is current added for mini flavor only as an example
and will probably change together with the other pieces such as the
main playbooks.

Change-Id: Ibe4851ca4bf57db075e1fc9eab9ca2e76f877753
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>