Update patch set 3
Patch Set 3:

Jenkins will fail for ubuntu because we don't use 'sudo -H' in the jjb job so the keystone roles fail because of it. For SUSE and CentOS, the HOME variable is set by default. As such, I would say that we need to force this patch in if the build is successful on CentOS and SUSE.

Patch-set: 3