Update patch set 1
Patch Set 1:

> How can I select the most recent set of files to download, without
 > knowing the exact name/version? Can this be a yum repo? Then I can
 > yum download and get the most recent. Or should we pin the build to
 > a particular version, and update the apex build script each time we
 > want to update the rpms?

Hi John
I was going to just have the one location where all the RPMs get placed every night. so you can just wget them from there. I wasn't planning on creating a repo. then when we freeze a release we can take a snapshot of the RPMs for a release... but I'm open to suggestions

Patch-set: 1