Build Timeout Wrapper for Yardstick
Yardstick builds are hanging and not being caught for 19+ hours. Using
the build timeout plugin we should be able to quickly stop these builds
and let the project know much sooner there is an issue with their
builds.

The default timeout is set to 15 minutes of no output from the console.
This timer is reset each time a line is output from the console. The
time can be increase in the future if there are operations in builds
which take longer than this.

ssh-agent-wrapper is included as it is the default from globals, and
defaults are superceded by fields added to jobs.

Change-Id: I9529ab7b6ab8819ceb91694fefbb4ecb3f5cf81c
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>