Comment Pending JJB Changes on Patchset
Each time a patchset is created that has an effect on Jenkins jobs, a
non-voting job will trigger that comments to Gerrit which jobs are
added, modified, or deleted.

This should help developers judge the overall impact an individual
patchset may have on Jenkins since JJB macros can be expanded to
any number of jobs.

Change-Id: I99f9b68bc7933abfa39a941d6e65a73f62e58e56
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>