Generalize Gerrit non-voting Review Macro Name
This patch updates the 'report-docs-build-result-to-gerrit' macro name
to reflect that it can work well for more than just the opnfvdocs project.

To take advantage of this macro a build merely needs to write
out the file 'gerrit_comment.txt' with the information they want posted
back to gerrit, and include the macro in the list of builders.

Change-Id: Ifbdc5734fe7e7231be0cfd16d6b8d69cb1383bfd
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>