Update patch set 3
Patch Set 3:

(1 comment)

I tried to implement it as generic as possible because for example 
functest-verify-x86-{stream} and functest-verify-arm-{stream} both match the functest-verify-{arch}-{stream} template and if an x86 job would be needed it would save some duplicate code or actually rewriting this section.
Also I used {installer} since fuel is or will not be the only option to be applicable.

Patch-set: 3