Update patch set 7
Patch Set 7:

posted my comments before reading Jose's comments but it would be good to get some kind of plan/indication about when the rename/change can be done. post danube.3.0, by milestone X etc. Tbh the earlier is better since renaming stuff will cause loss of build history and it might cause unnecessary trouble for you to explain things to community.

I'm aware of the need to be in a hurry and will not block this if you can provide satisfactory answer as both me and Jose seem to have same concern.

Patch-set: 7