Update patch set 8
Patch Set 8:

In CI runs only, Functest tries to fetch some information from the deployment. To do that, it needs to know which installer it is with the environment variable INSTALLER_TYPE. If you tell Functest the deployment is with Fuel, it will try to access the default ip/password 10.20.0.2/r00tme to do some things like 'fuel node'. 

I wouldn't like to change this behavior in Functest as it works well so far. 

This mechanism is also used by some tests cases like SFC, which need to access the deployment.

And as well, I think it doesn't make sense to create confusion. If the installer is called MCP, let's use that name in the jobs and scripts instead of creating if-conditions with the branch names...

Patch-set: 8