Directly Run Anteater Docker Container
The current approach is to run /bin/bash in a fully privilaged docker
container as the root user and exec the anteater command from this.

There are a couple of reasons this approach doesn't make sense:

1) anteater is not a long running service
2) anteater doesn't need any privilaged access to the host
3) anteater is already a compiled binary and can be ran directly

Instead of violating the priciple of least privilage, the
releng-anteater docker container should be ran as a non-root user.
Because this requires an existing user account (/etc/passwd) inside the
container, the simplest approach is to use 'nobody'. Once a non-root
user has been added to the container, this script should be updated to
use that user instead.

This should keep files from being written to the workspace as the root
user.

Note: '--rm' is added as well to ensure volumes do not persist past the
container lifecycle and lead to build server running out of disk space.

JIRA: RELENG-250

Change-Id: I1ec90b3737abf591b6b3373fe2fc8f52cdcfb11a
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>