Update patch set 1
Patch Set 1:

I did test this locally, and it worked. I was surprised to find it didn't run on ericsson-build3. Since the error was with the docker run command, I believe the daemon logs will have to be examined to find out why it failed.

Sudo shouldn't be needed in the Dockerfile unless a RUN command that requires root permissions comes after a USER command.

Patch-set: 1