Yamllint Job for Octopus, Pharos, and Anteater
Pharos, Octopus and Securedlab are listed in the wrong lint check job
so new patchsets aren't being ran against yamllint.

Anteater is included in the list of yamllint checks due to it being
part of Infra-WG and containing yaml files.

Fixes still need to be done for the securedlab and releng repos before they
are added at a later time.

JIRA: INFRA-101

Change-Id: I696b233e076899107180ed9bbad9a7f7785058d3
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>