Bifrost+OSA install on baremetal with config file generation
I publish this procedure and set of ansible scripts to share the status of
pseudo-installer on Orange pods. All configs needed by bifrost and osa are
generated based on pdf-like files (see /vars), this could move to real pdf
once the install will our pods will be finalized. Actual generated files are:
- bifrost inventory
- node network setup, based on node role
- ansible inventory for node network setup
- ssh config
- osa config
As it is not an installer, there is no all-in-one script, but it can contains
good material for xci on baremetal. If you think it is a good basis I can
work on a pdf based version.

Change-Id: I577cb11062fc07a15961184e49ffce18685d4a64
Signed-off-by: Blaisonneau David <david.blaisonneau@orange.com>