Update patch set 2
Patch Set 2: Code-Review-1

It seems this is not a solution. What doctor team is expecting, is to run doctor CI jobs on one of virtual pods in OOL or ZTE, for load balancing testing jobs. We don't want to run the same jobs in new ZTE pods.

Patch-set: 2
Label: Code-Review=-1