Update patch set 1
Patch Set 1:

flake8 verification is not part of this job, it's another one which is triggered anyways.

Is there any option to exclude branches? Something like regex 

    branch: !stable/danube

That would also solve the problem but I am not sure if that is possible...

We can abandon this and do manual verify +1 in the patches in stable/danube or merge this and remove this condition again when we branch for Euphrates.

Patch-set: 1