Update patch set 23
Patch Set 23:

(1 comment)

@Morgan: 
There are lots of issues in releng which have been fixed in Functest. Please add me as reviewer to topic about requirements, docker and tox.

I think we have to decide here which versions to install. I think it must be updated to follow OpenStack requirements.

Patch-set: 23