Update patch set 1
Patch Set 1:

(2 comments)

If the purpose of this patch is to pass a milestone (healthcheck working), then you don't need functest-suite. By running the daily job and verifying that healthcheck is passing is enough.  

If the purpose is to use the healthcheck as a single jjob to verify gerrit patches (gates) as Apex team is doing, then it makes sense to have it. Otherwise, it is not really needed.

Patch-set: 1
Reviewer: Gerrit User 459 <459@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0