Update patch set 3
Patch Set 3:

> as new features are added, I would expect associated unit tests to
 > be ensure coverage and stability of the overall code

As regards to this point, I am the one should be blamed, I was hurry up with working out a framework before Beijing Summit, so there's no unit tests for the new codes. So considering this historic reason, I would suggest to merge it ala the implementation is fine. 
what's your opinion?

Patch-set: 3