Multiple Dockerfile Builds per Project
Instead of defining docker build jobs in terms of projects, we define
them in terms of dockerrepos. Each dockerrepo is associated with a
specific project and may have a seperate Dockerfile (which defaults to
'Dockerfile' if unspecified).

This should allow for running multiple docker builds out of a single
project repository without having to modify the current docker build
scripts.

The output of this job configuration should match what is currently in
Jenkins.

Change-Id: I9089d57a0c7cc56be4225b56e96bec56be258f1f
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>