Add support for review jupyter notebook from nbviewer
It is done by uploading the notebook under reviewer to artifacts
and add the nbviewer link in gerrit comments

It is needed because reviewing jupyter notebook[1] directly in
gerrit is very difficult. See example in [2]. It will be much
easier if we can preview the result in nbviewer. See example in [3]

[1]: http://jupyter.org/
[2]: https://gerrit.opnfv.org/gerrit/#/c/37227/
[3]: https://nbviewer.jupyter.org/urls/git.opnfv.org/qtip/plain/examples/storage-qpi-report/storage-qpi.ipynb

Change-Id: I3e4c301f362ff6ac97aebc9d09c7103278efb136
Signed-off-by: Yujun Zhang <zhang.yujunz@zte.com.cn>