Update patch set 3
Patch Set 3:

> (1 comment)

I wouldn't agree with this, take Refstack for example, it doesn't expose role info to user as well.
From dovetail's requirements, we can see finally vendor/company may will be introduced to manage the authentication, that case how to define a role of a user specifically?
And as far as the common TestAPI usage is concerned, different projects/scenarios/results... may mean different roles to the same user. For example, A is the PTL of Functest, but meanwhile he is a contributor in Dovetail, it means, he is the owner of Functest but only is a user of Dovetail, that way how to define the user's role?

Patch-set: 3