Update patch set 2
Patch Set 2:

> no, it's not required by OPNFV, but if we start using it everywhere
 > maybe it can become part of the "best practices" to be inline with
 > OpenStack.

If we want to follow OpenStack, we'd better migrate the check mechanism as well, because that global-requirements.txt updates day-by-day, and check each project periodically, once an inconsistent detected, it will raise an error. if we don't follow the whole procedure, our requirements will be out-of-date eventually.

Patch-set: 2
Label: Verified=0