Update patch set 5
Patch Set 5:

Sorry https://gerrit.opnfv.org/gerrit/#/c/37243/ and https://gerrit.opnfv.org/gerrit/#/c/37335/ should be submitted together. but to some merge order reason 37243 is not merged last night. you can rebase 37243 to see whether the patch pass the jenkins-ci or not.
And I see lots of other irrelevant things done in this patch, I would recommend to remove them. only one thing in a patch is regulated in OpenStack, and in Functest, we were asked to follow the rule.

Patch-set: 5