Update patch set 1
Patch Set 1:

> (1 comment)

Use this solution, maybe we still need to create the folder functest/ci, and the download the script in the dir. But i think the workaroud is so weird.

To Ryota:
I have a question, why we use functest builder for our verify jobs.
I think we just need to use 'fetch_OS_creds.sh' to get the credentials and then run our jobs. What do you think?

Patch-set: 1