Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

The idea is good according to the latest discussions about each project hosting their build scripts.
I would like to have more flexibility when it comes to the path and the filename of the build script. 

Not sure if verify docker is needed.

Patch-set: 2
Label: Code-Review=-1