Update patch set 1
Patch Set 1:

> (1 comment)

As we talked about in the log.  It looks like with multijob it sets the current status before the multijob phases are executed.  I'm not sure how to work around that so I think we can try a publisher instead which will execute post job.  Can you please install the groovy postbuild plugin?

Patch-set: 1
Label: Verified=0