Update patch set 3
Patch Set 3:

(1 comment)

Micahel, the daily job looks good!

Just looking for some clarification on the 'why' for this patchset, which should be reflected in the commit message.

Is there a specific need to run a deployment @midnight? If the build is expected to happen before the deployment (assuming that's why the timed trigger is set to 11PM), then you may want to not use H in the minute field, as this could possible schedule build at '59 23 * * *'.

I'm guessing the build is currently a place holder until a build script can be written?

Patch-set: 3