Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

The patch is mixing the concepts of PATH and remote repository, which is confusing.

It would be good if this could be extended for a generic solution that could be applied to any dependency in ansible-role-requirements (e.g. nova, keystone), not specific to neutron.

Patch-set: 1
Label: Code-Review=-1