fuel, armband: Temporarily remove pxebr deploy arg
Fuel and implicitly Armband switched to a newer format for parsing
the "-B" deploy arg, used to pass bridge(s) information.

For now, we will default to ",,,", which is equivalent to not
setting any default value for any of the 4 possible bridges.

NOTE: Armband overrides this on a per-pod basis, since bridge
information is POD specific in our lab.

Change-Id: I3635b0eb6f8233cf99022a97d5642f2e49197e3b
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>