fuel, armband: Drop "-H" arg, fix bridge default
MCP does not have a self-test healtcheck suite, similar to old
Fuel, so we dropped the "-H" deploy argument (which used to disable
healtcheck autorun).
Adjust deploy command to reflect this installer change.

While at it, fix leftover "pxebr" default installer parameter,
which we missed in the previous commit where we changed the
default only in releng deploy scripts.

Change-Id: I55e1f5cad17687bdf304fbf04d67366dd32ed42a
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>