Update patch set 1
Patch Set 1:

(4 comments)

Hey Alec,

From what I can tell from this patch you want to add a job triggered by the comment 'buildvm' which kicks off a build of the nfvbench.

I have a couple of suggestions in the comments that should hopefully help you get there, and those related to job naming are to help align with our CI Evolution[1] changes.

If you have a moment to locally yamllint[2] this as well that would be awesome!

[1] https://wiki.opnfv.org/display/INF/CI+Evolution
[2] https://pypi.python.org/pypi/yamllint

Patch-set: 1
Label: Verified=0