Apex: fix groovy build result update
The groovy script is working, but it is changing the 'apex test suite'
phase to unstable and we really want it to change the 'scenario' job
result to unstable if deploy passed but tests failed.  Therefore this
patch moves the groovy plugin to execute on the scenario job, and looks
for deploy successful before changing the build result.

Change-Id: I1b6907f4fb2647e644c4a627be6e481fff32f195
Signed-off-by: Tim Rozet <trozet@redhat.com>