Update patch set 1
Patch Set 1:

(2 comments)

In my build I'll ultimately need to build 2 artifacts:
a VM (using that build_vm.sh script) and a docker container (will get to that next).
And they both have to be triggered by tag. 
Upon gerrit review, the artifacts must be built if the appropriate tag is pushed
Upon merge, the artifacts will need to be pushed to docker hub and the opnfv artifact repo
Is that possible and how?

Patch-set: 1
Label: Verified=0