Update patch set 7
Patch Set 7:

I agree with some of Cedric's comments:

This is supposed to be part of the Gerrit Merge and Verify steps, so it really should be under ci/docker-verify and ci/docker-merge-etc...

My plan is to use this job to create a local container (never pushed to docker hub, ie: COPY .) to verify that the newly proposed patch does not break the ability to create the compose containers.

I think just a simply docker-compose of local content (COPY .) and a quick API get (GET http//127.0.0.1:StorPerf-Port/API) to verify that it still responds is all I want.

This should be run as part of the verify job, in parallel to the existing ci/verify.sh job.  Alternatively, I could spin up the container as part of the regular ci/verify.sh, but then verify.sh checks two things; 

1) the unit tests pass, and
2) the composed container still starts up.

With this change (and changing it back to ci/verify-container.sh) I can run both in parallel and have independent +1/-1 depending on what passed/failed

Patch-set: 7