Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

I think this can be further simplified by checking GERRIT_EVENT_COMMENT_TEXT. I think this should be done in global-jjb though so that we can deploy it to all the projects. See RELENG-373.

I also think having the reviewer specify the specific job they want to test in the sandbox to be a better idea. They can pass wildcards too if we escape the string appropriately in the shell script.

Patch-set: 1
Label: Code-Review=-1