jjb: xci: bifrost-verify: Disable errexit when uploading logs
The errexit option is there to catch failures when the job is executed.
We don't want to mark the entire job as failed when we have troubles
uploading the logs when the actual job has passed so we disable the
errexit option to make the whole process less aggressive. Moreover, we
move the build log upload to the end of the operation so we can catch
as much output as possible.

Change-Id: Ice126471a8c837c54c95c8ef0935c059d37ea91b
Signed-off-by: Markos Chandras <mchandras@suse.de>