Apex: Fixes bug with blocking
We have blocking so that deploy wont run while another deploy or
functest is running.  However there is an issue where after a deploy
completes, the next in line deploy starts at the same time as the
functest starts for the previous deploy.  This patch adds a quiet period
to the deploy job of 30 seconds, which hopefully will offset the start
time between the two and resume correct blocking.

Change-Id: Id94d075f2d2e48952a5c2186eeb81d09bb68c49b
Signed-off-by: Tim Rozet <trozet@redhat.com>