Update patch set 2
Patch Set 2:

My initial reaction to this is -1, but after reading through some of the context in 'install.sh' I can understand this is a complex issue involving the use of PBR and data-files.

Sorry I don't have enough time right now to dive into the internals of setuptools/pbr, and since I'm assuming you've tested this works and it fixes the problem, +1.

Patch-set: 2