Update patch set 2
Patch Set 2:

One last note: Since this is at the top of setup.py, it will be ran every time setup.py is invoked, which may not be what you want. Since I only see calls to 'setup.py install' this may not be an issue, but you might need to have the subprocess call only happen _if_ 'setup.py install' is called, and not say 'setup.py build'.

https://stackoverflow.com/questions/17806485/execute-a-python-script-post-install-using-distutils-setuptools#18159969

Patch-set: 2