Apex, Yardstick: Fixes fetching creds
There is a bug where by default the INSTALLER_IP is set to 192.168.X.X
and in yardstick the IP is never found for Apex before trying to fetch,
so fetching fails to an invalid IP address.

This patch makes fetch os creds more robust to check if the INSTALLER_IP
is valid and if not, then attempt to find the correct IP for Apex.

Change-Id: Ie18d6e70f77ccff3592fb37c6d0b6e87b6435cf8
Signed-off-by: Tim Rozet <trozet@redhat.com>