functest: verify: Hack around update_job issue
It seems that update_job cannot live-update a job that used not
to be multijob to multijob.

Instead of trying to force this, rename our functest-verify-* jobs
to something else, deleting old fuel-verify-* jobs completely and
introducing a new naming.

Once the new naming is in effect, this commit should be reverted,
achieving the not-so-live-update of fuel-verify-* jobs.

Change-Id: Ic99ca7f4be17105bef28420c130fe68ccdbe064d
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>