Update patch set 2
Patch Set 2:

Afaik, **/* would trigger a build only if files in subdirs are changed, but not if the root-dir files are touched (e.g. INFO, LICENSE etc.).
As for disable-strict-forbidden-file-verification=true without 'file-paths', my understading was that it allows builds to be triggered if any file that is not forbidden was touched, even if forbidden files were touched too; the default behavior being no build will be triggered when a forbidden file is touched ...

Patch-set: 2
Label: Code-Review=0