Apex: Explicitly define file paths
Due to bug with forbidden file paths conflicting with file paths this
explicitly defines all the file paths in apex.  This isn't the greatest
solution in the case where someone adds a new folder, but chances are
they will also modify one of the other files as well.  Note it also
fixes a bug where unit tests woudl run twice because it would be
triggered as part of apex-verify-master and apex-unit-tests-master.

Change-Id: I6b10f8eb70c2cd7c8dd5c01363e06b7458c00127
Signed-off-by: Tim Rozet <trozet@redhat.com>