Switch daily job to Alpine containers
The following alpine containers will be launched one after the others:
 - healthcheck
 - smoke
 - features
 - vnf

Old container is kept for suite (which could be refactored later)

Jenkins exit condition (return_value) and push_test_logs to be checkeded.

Regarding the return condition, I did not reapply the mechanism that returns false
when one of the test suite is FAIL. This was misleading for lots of people as
for them, red means functest is FAIL.The argument was, if functest jenkins status
is blue, people believe that tests are PASS and do not consider the results, which
may be FAIL, which may be also misleading...

Due to the refactoring I think we can be more Functest Centric and consider
that if we succeeded in running the test, our CI status shall be PASS (same
behavior than other test projects..HA test FAIL in yardstick do not prevent blue
 status).

Scenario owner must have a look at the reportign page to get more details.
Functest PASS means that tests have been succesfully launched, so they know that
there are available results
To summarize
previously Functest OK but 1 or more tests FAIL => red status in Jenkins
currently Functest OK but 1 or more tests FAIL => blue status in Jenkins

Regarding the push to artifacts, it should be transparent (volume) but the last
log links listed on huawei-virtual5 are broken [1]

[1]: https://build.opnfv.org/ci/view/functest/job/functest-alpine-compass-huawei-virtual5-daily-master/27/consoleFull

Change-Id: I4ed036dda3daae3d46726e5946f9d04897bab9bb
Signed-off-by: Morgan Richomme <morgan.richomme@orange.com>