Apply aarch64 patch for Alpine functest containers
A single patch for all alpine containers is applied and it is
found in /docker/core path, under Dockerfile.aarch64.patch
name. The patch is no longer applied with Dockerfile.aarch64
file as result. The original Dockerfile file is being replaced.
The same applies for the old Dockerfile.arch.patch, for ubuntu
image containers, which has not yet been removed.
The changes is applying both patches (for ubuntu and alpine image).

It depends on https://gerrit.opnfv.org/gerrit/#/c/41677/
JIRA: ARMBAND-310

Change-Id: I2efa1b58c4b82a21eaedb70b922929a35c1f96b8
Signed-off-by: Delia Popescu <delia.popescu@enea.com>