xci: Make it possible to skip deployment and healthcheck
This change adds possibility for patch submitters to skip the full
xci-verify if the change doesn't impact the deployment.

Since the patch will be reviewed by others, reviewers can always
ask the submitter to run it through full xci-verify by changing
the topic and posting reverify or recheck as comment to gerrit
change.

The keyword to put in topic branch is skip-verify.

Change-Id: Ib7be71cadd1e1ddff181936ae77753fbfbf3a1d3
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>