xci: Make it possible to skip deployment and healthcheck
This change adds possibility for patch submitters to skip the full
xci-verify if the change doesn't impact the deployment.

Since the patch will be reviewed by others, reviewers can always
ask the submitter to run it through full xci-verify by amending
the commit message.

The keyword to put in commit message is skip-xci-verify.

Change-Id: Ib7be71cadd1e1ddff181936ae77753fbfbf3a1d3
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>