Update patch set 1
Patch Set 1: Code-Review-1

You dont need the build.  The deploy should pull the latest danube rpms and install.  The issue I see is in this log:
https://build.opnfv.org/ci/job/apex-deploy-baremetal-danube/635/console

We have renamed opnfv-apex-common to python34-opnfv-apex in master/euphrates so it is failing.  We will need to add back legacy support in releng to detect in the deploy if it is danube and then use the old package names:
https://github.com/opnfv/releng/blob/master/jjb/apex/apex-download-artifact.sh#L42

Patch-set: 1
Reviewer: Gerrit User 334 <334@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1