Update patch set 1
Patch Set 1:

> (1 comment)

Hi, Julien,
Since check-jinja2.sh will be moved to Pharos git repo, adding it to the list of monitored files here won't work.
Instead, we could try monitoring the pharos git submodule, but even that will only trigger a check when the pharos git submodule reference is updated. Since we use `git submodule update --remote`, we will always get pharos git branch HEAD, even though the submodule reference was not updated in securedlab.
So, this is tricky, if you have a suggestion on how to approach this, I'd love to hear it.

Patch-set: 1