Apex: Fixes iso verify depending on previous build workspace
There is a bug where if there are multiple builds queued and daily build
completes, the following iso verify job will try to use the workspace of
the completed daily build to get the iso file.  However, if another
build job has already started it may clean and overwrite the workspace
of the daily build job, and the iso verify job will fail because the
file is now gone.  This makes the build job copy the iso to a tmp
directory for apex iso verify to consume.

This should be safe since only one daily can run at a time on the host
and daily jobs build and iso verify have to always execute on the same
node.

Change-Id: Ie8e32c4abefbc311e505688d6da2b26ae08ed98f
Signed-off-by: Tim Rozet <trozet@redhat.com>