Update patch set 4
Patch Set 4:

> I think it is still fuel connected because the host_prefix variable
 > is something which does not make sense in xci but does in fuel when
 > having several deployments. We should figure out a way to identify
 > the node where the logs are being fetched which works for all
 > installers

IMHO I think it does make sense in all installers because if you have a HA deployment want to know for instance from which node each ovs-table ccomes from.

I think this is the purpose of the host_prefix variable to indentify the hosts and not the fuel env only.

So when we read the ovs-logs we can understand instantly in which host this log correspond

Patch-set: 4