jjb: xci: Create XCI test script
We need to respect the environment during testing so we need to create
a proper test script which exports all the variables that matter for our
tests. We do that before we create the VM so the VM creation script can
rsync the test for us. We also properly escape the DISTRO variable when
destroying the VM.

Change-Id: Ic3fff40868155af31762e458c1faee24354c0712
Signed-off-by: Markos Chandras <mchandras@suse.de>